+++ modelopt/torch/export/plugins/mcore_llama.py
@@ -101,7 +101,7 @@ eagle_llama_causal_lm_export: dict[str, CustomModuleMapping] = {
     "parallel_draft_heads.medusa_layers": NameRemapping(
         "parallel_draft_heads.{}.medusa_layers.{}.linear."
     ),
-    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.{}.lm_head."),
+    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.lm_head."),
 }
 
 eagle3_llama_causal_lm_export: dict[str, CustomModuleMapping] = {
@@ -120,7 +120,7 @@ eagle3_llama_causal_lm_export: dict[str, CustomModuleMapping] = {
     "parallel_draft_heads.medusa_layers": NameRemapping(
         "parallel_draft_heads.{}.medusa_layers.{}.linear."
     ),
-    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.{}.lm_head."),
+    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.lm_head."),
 }
 
 eagle3_deep_llama_causal_lm_export: dict[str, CustomModuleMapping] = {
@@ -140,7 +140,7 @@ eagle3_deep_llama_causal_lm_export: dict[str, CustomModuleMapping] = {
     "parallel_draft_heads.medusa_layers": NameRemapping(
         "parallel_draft_heads.{}.medusa_layers.{}.linear."
     ),
-    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.{}.lm_head."),
+    "parallel_draft_heads.lm_head": NameRemapping("parallel_draft_heads.lm_head."),
 }