+++ modelopt/torch/export/unified_export_megatron.py
@@ -1069,12 +1069,12 @@ class GPTModelExporter:
 
         parallel_draft_heads = getattr(eagle_module, "parallel_draft_heads", None)
         if parallel_draft_heads is not None:
-            for head_id, head in enumerate(parallel_draft_heads):
-                self.rules["parallel_draft_heads.lm_head"](head.lm_head, head_id)
-                for layer_id, layer in enumerate(head.medusa_layers):
+            for head_id, head in enumerate(parallel_draft_heads.medusa_heads):
+                for layer_id, layer in enumerate(head):
                     self.rules["parallel_draft_heads.medusa_layers"](
                         layer.linear, head_id, layer_id
                     )
+            self.rules["parallel_draft_heads.lm_head"](parallel_draft_heads.lm_head)
 
     def _get_state_dict(self):
         model = self.model